requirements.txt
Flask==2.2.5
requests==2.31.0
beautifulsoup4==4.12.2
mysql-connector-python==8.2.0
gunicorn==21.2.0  # Optional, for production deployment
